haywire will now keep allocating memory until the request is completely parsed.
Previously, we were assuming that the alloc callback was only called once per request and that the read callback was called when all the data was available.
Unfortunately, that's not the case, and the read callback may be called when only a fraction of the incoming data is available, especially under high concurrency.

To work around this, we allocate/reallocate the request buffer as required to keep up with the incoming data.
Since the buffer may be reallocated to a different memory region, we save the offset for the URL, headers and the body in the original buffer and then repoint all of them when we finish parsing the request, so that they're in the same position in relation to the beginning of the final request buffer.

+++ include/haywire.h
@@ -150,7 +150,7 @@ typedef struct
     char* parser;
     bool tcp_nodelay;
 } configuration;
-
+    
 typedef struct
 {
     unsigned short http_major;
@@ -160,7 +160,7 @@ typedef struct
     hw_string* url;
     void* headers;
     hw_string* body;
-    int body_length;
+    size_t body_length;
 } http_request;
 
 typedef	void* hw_http_response;
